create file for data fetching
