+++ src/__tests__/App.test.js
@@ -1,6 +1,6 @@
 import { render, screen, waitForElementToBeRemoved } from '@testing-library/react';
 import MockAdapter from "axios-mock-adapter";
-import api from '../api';
+import api from '../api/api';
 import App from '../App';
 
 describe('<App />', () => {
